# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tutor Reservation API",
        "description": "RESTful API for tutor reservations",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Main server"
        }
    ],
    "paths": {
...
        "/api/tutors": {
            "get": {
                "tags": [
                    "Tutors"
                ],
                "summary": "Get tutors",
                "operationId": "9bbbb366948632e3ea8de58ac09a149a",
                "responses": {
                    "200": {
                        "description": "Tutors retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tutors/{id}": {
            "get": {
                "tags": [
                    "Tutors"
                ],
                "summary": "Get tutor by ID",
                "description": "Returns the tutor information and reservation statistics.",
                "operationId": "0503cd6af448456d5f8e33650b1b1f7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tutor ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 26
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tutor retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Tutor not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tutor/profile": {
            "get": {
                "tags": [
                    "Tutors"
                ],
                "summary": "Get authenticated tutor profile",
                "operationId": "a5e2980c07d917c998f3093ee5a4d8a5",
                "responses": {
                    "200": {
                        "description": "Tutor profile retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Authenticated user is not a tutor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tutors"
                ],
                "summary": "Update authenticated tutor profile",
                "operationId": "e78801302691c9816e665290f2fc20f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_names": {
                                        "type": "string",
                                        "example": "Carlos"
                                    },
                                    "last_names": {
                                        "type": "string",
                                        "example": "Perez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos.perez@test.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "3001234567",
                                        "nullable": true
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": "photos/tutors/carlos.jpg",
                                        "nullable": true
                                    },
                                    "hourly_rate": {
                                        "type": "integer",
                                        "example": 40000,
                                        "maximum": 100000,
                                        "minimum": 30000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tutor profile updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Authenticated user is not a tutor"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tutors/{id}/calendar": {
            "get": {
                "tags": [
                    "Tutors"
                ],
                "summary": "Get tutor calendar availability",
                "operationId": "020af2e05f6473e1b4ef9be3e844c650",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tutor ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 26
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date to check availability",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-08-18"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tutor calendar retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Tutor not found"
                    },
                    "422": {
                        "description": "Invalid date"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tutors/reservations": {
            "get": {
                "tags": [
                    "Tutors"
                ],
                "summary": "Get authenticated tutor reservations",
                "description": "Returns all reservations belonging to the authenticated tutor, including the student who made each reservation.",
                "operationId": "db974600d3a8fbf01d0c923ce0d6ef3c",
                "responses": {
                    "200": {
                        "description": "Tutor reservations retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Authenticated user is not a tutor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tutors/calendar": {
            "get": {
                "tags": [
                    "Tutors"
                ],
                "summary": "Get authenticated tutor calendar",
                "description": "Returns the availability calendar of the authenticated tutor.",
                "operationId": "a2285e7020ef68eed19b1d7a3dda9e80",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date to check availability",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-08-18"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tutor calendar retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Authenticated user is not a tutor"
                    },
                    "422": {
                        "description": "Invalid date"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tutors"
                ],
                "summary": "Set authenticated tutor availability",
                "description": "Replaces the authenticated tutor availability for a date. Hours must be whole-hour slots between 08:00 and 17:00.",
                "operationId": "8347f5ddf17cc9361c8abff80f57ec4c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "date",
                                    "hours"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Horario de clases de septiembre",
                                        "maxLength": 150
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-09-20"
                                    },
                                    "hours": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "pattern": "^([0-1][0-9]):00$",
                                            "example": "08:00"
                                        },
                                        "maxItems": 10,
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tutor availability updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Authenticated user is not a tutor"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Bearer token generated by Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}